backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
rembg
onnxruntime
Pillow
